xerate/requirements.txt
requests
google-cloud-secret-manager
google-cloud-storage
google-api-core

